+++ src/pages/Details.js
